+++ hexa/pipelines/views.py
@@ -9,6 +9,7 @@ from django.shortcuts import get_object_or_404, redirect, render
 from django.utils.translation import ugettext_lazy as _
 from django.views.decorators.http import require_http_methods
 
+from hexa.metrics.decorators import do_not_track
 from hexa.pipelines.datagrids import DAGRunGrid, EnvironmentGrid
 from hexa.pipelines.models import Environment, Index
 from hexa.plugins.connector_airflow.models import AirflowAPIError, DAGRun
@@ -39,6 +40,7 @@ def index(request: HttpRequest) -> HttpResponse:
     )
 
 
+@do_not_track
 def index_refresh(request: HttpRequest) -> HttpResponse:
     dag_runs = DAGRun.objects.filter_for_user(request.user)
     for run in dag_runs.filter_for_refresh():