+++ hexa/plugins/connector_s3/views.py
@@ -6,6 +6,7 @@ from django.shortcuts import get_object_or_404, redirect, render
 from django.urls import reverse
 from django.utils.translation import ugettext_lazy as _
 
+from ...metrics.decorators import do_not_track
 from .api import generate_download_url, generate_upload_url
 from .datacards import BucketCard, ObjectCard
 from .datagrids import ObjectGrid
@@ -148,7 +149,8 @@ def object_upload(request, bucket_id):
     return HttpResponse(upload_url, status=201)
 
 
-def object_refresh(request, bucket_id):
+@do_not_track
+def bucket_refresh(request, bucket_id):
     bucket = get_object_or_404(
         Bucket.objects.filter_for_user(request.user), pk=bucket_id
     )